src/main/webapp/app/entities/warning-rule/route/warning-rule-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IWarningRule, WarningRule } from '../warning-rule.model';
import { WarningRuleService } from '../service/warning-rule.service';

import { WarningRuleRoutingResolveService } from './warning-rule-routing-resolve.service';

describe('WarningRule routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: WarningRuleRoutingResolveService;
  let service: WarningRuleService;
  let resultWarningRule: IWarningRule | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(WarningRuleRoutingResolveService);
    service = TestBed.inject(WarningRuleService);
    resultWarningRule = undefined;
  });

  describe('resolve', () => {
    it('should return IWarningRule returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultWarningRule = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultWarningRule).toEqual({ id: 123 });
    });

    it('should return new IWarningRule if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultWarningRule = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultWarningRule).toEqual(new WarningRule());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as WarningRule })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultWarningRule = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultWarningRule).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
